LayoutTests/inspector/dom-debugger/dom-breakpoints-expected.txt

Tests for DOM breakpoints.


== Running test suite: DOMBreakpoints
-- Running test case: BasicBreakpoint
PASS: Added 'subtree-modified' breakpoint.
PASS: Breakpoint should not be disabled.
PASS: Breakpoint should have node identifier.
PASS: Breakpoint URL should match document URL.
-- Running test teardown.

-- Running test case: DOMBreakpoints.SubtreeModified.BreakpointEnabled
PASS: Added 'subtree-modified' breakpoint.
PASS: Breakpoint should have expected type.
Call DOM operation.
PAUSED:
PASS: Pause reason should be DOM.
CALL STACK:
0: [F] subtreeModifiedTest
1: [P] Global Code
-- Running test teardown.

-- Running test case: DOMBreakpoints.SubtreeModified.BreakpointDisabled
PASS: Added 'subtree-modified' breakpoint.
Wait for evaluate in page to return.
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMBreakpoints.SubtreeModified.DebuggerDisabled
PASS: Added 'subtree-modified' breakpoint.
Wait for evaluate in page to return.
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMBreakpoints.SubtreeModified.RemoveBreakpoint
PASS: Added 'subtree-modified' breakpoint.
Remove breakpoint.
Wait for evaluate in page to return.
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMBreakpoints.AttributeModified.BreakpointEnabled
PASS: Added 'attribute-modified' breakpoint.
PASS: Breakpoint should have expected type.
Call DOM operation.
PAUSED:
PASS: Pause reason should be DOM.
CALL STACK:
0: [F] attributeModifiedTest
1: [P] Global Code
-- Running test teardown.

-- Running test case: DOMBreakpoints.AttributeModified.BreakpointDisabled
PASS: Added 'attribute-modified' breakpoint.
Wait for evaluate in page to return.
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMBreakpoints.AttributeModified.DebuggerDisabled
PASS: Added 'attribute-modified' breakpoint.
Wait for evaluate in page to return.
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMBreakpoints.AttributeModified.RemoveBreakpoint
PASS: Added 'attribute-modified' breakpoint.
Remove breakpoint.
Wait for evaluate in page to return.
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMBreakpoints.NodeRemoved.BreakpointEnabled
PASS: Added 'node-removed' breakpoint.
PASS: Breakpoint should have expected type.
Call DOM operation.
PAUSED:
PASS: Pause reason should be DOM.
CALL STACK:
0: [F] nodeRemovedTest
1: [P] Global Code
-- Running test teardown.

-- Running test case: DOMBreakpoints.NodeRemoved.BreakpointDisabled
PASS: Added 'node-removed' breakpoint.
Wait for evaluate in page to return.
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMBreakpoints.NodeRemoved.DebuggerDisabled
PASS: Added 'node-removed' breakpoint.
Wait for evaluate in page to return.
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMBreakpoints.NodeRemoved.RemoveBreakpoint
PASS: Added 'node-removed' breakpoint.
Remove breakpoint.
Wait for evaluate in page to return.
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: RemoveAllBreakpointsForNode
PASS: Added 'subtree-modified' breakpoint.
PASS: Added 'attribute-modified' breakpoint.
PASS: Added 'node-removed' breakpoint.
PASS: Removed 3 breakpoints.
PASS: DOM node should have no breakpoints.
-- Running test teardown.

-- Running test case: SetBreakpointWithInvalidNodeId
Attempting to set breakpoint.
Protocol result: Missing node for given nodeId
PASS: Protocol should return an error.
-- Running test teardown.

-- Running test case: SetBreakpointWithInvalidType
Attempting to set breakpoint.
Protocol result: Unknown DOM breakpoint type: custom-breakpoint-type
PASS: Protocol should return an error.
-- Running test teardown.

-- Running test case: RemoveBreakpointWithInvalidNodeId
Attempting to remove breakpoint.
Protocol result: Missing node for given nodeId
PASS: Protocol should return an error.
-- Running test teardown.

-- Running test case: RemoveBreakpointWithInvalidType
Attempting to remove breakpoint.
Protocol result: Unknown DOM breakpoint type: custom-breakpoint-type
PASS: Protocol should return an error.
-- Running test teardown.
